+++ CHANGELOG.md
@@ -19,6 +19,7 @@ _Internal changes, like typos or internal fixes to the json-schema are not liste
 
 * New **GDDType**: `color-rrggbbaa` [PR #16](https://github.com/SuperFlyTV/GraphicsDataDefinition/pull/16)
 * Version 1.0.0 of the [javascript npm library](https://www.npmjs.com/package/graphics-data-definition) is released.
+* Optional properties `authorName` and `authorEmail` added, [PR #17](https://github.com/SuperFlyTV/GraphicsDataDefinition/pull/17)
 
 ## 2022-10-28